refactor(omegaform): delete OmegaFormStuff.ts after full extraction
