+++ packages/vue-components/__tests__/OmegaForm/TaggedUnionNested.test.ts
@@ -1,7 +1,7 @@
 // __tests__/OmegaForm/TaggedUnionNested.test.ts
 import { S } from "effect-app"
 import { describe, expect, it } from "vitest"
-import { generateMetaFromSchema } from "../../src/components/OmegaForm/OmegaFormStuff"
+import { generateMetaFromSchema } from "../../src/components/OmegaForm"
 
 const schema = S.Struct({
   aString: S.NonEmptyString255,
