+++ packages/vue-components/__tests__/OmegaForm/ValidationLocalization.test.ts
@@ -1,6 +1,6 @@
 import { S } from "effect-app"
 import { describe, expect, it } from "vitest"
-import { generateMetaFromSchema } from "../../src/components/OmegaForm/OmegaFormStuff"
+import { generateMetaFromSchema } from "../../src/components/OmegaForm"
 
 describe("S.Email format detection", () => {
   it("flags S.Email fields with format: 'email' on their meta", () => {
