+++ packages/vue-components/__tests__/test-array-meta.test.ts
@@ -1,6 +1,6 @@
 import { S } from "effect-app"
 import { describe, expect, it } from "vitest"
-import { generateMetaFromSchema } from "../src/components/OmegaForm/OmegaFormStuff"
+import { generateMetaFromSchema } from "../src/components/OmegaForm"
 
 const schema = S.Struct({
   a: S.Array(S.Struct({
