+++ packages/vue-components/src/components/OmegaForm/OmegaAutoGen.vue
@@ -23,7 +23,8 @@
 import { type DeepKeys } from "@tanstack/vue-form"
 import { Order } from "effect-app"
 import { computed } from "vue"
-import { type FieldMeta, type FieldPath, type OmegaAutoGenMeta, type OmegaInputProps } from "./OmegaFormStuff"
+import { type FieldMeta } from "./meta/types"
+import { type FieldPath, type OmegaAutoGenMeta, type OmegaInputProps } from "./types"
 
 type NewMeta = OmegaAutoGenMeta<From, To, Name>
