+++ packages/vue-components/src/components/OmegaForm/OmegaInputVuetify.vue
@@ -207,7 +207,7 @@
   generic="From extends Record<PropertyKey, any>, Name extends DeepKeys<From>"
 >
 import { type DeepKeys } from "@tanstack/vue-form"
-import { getInputType } from "../OmegaForm/OmegaFormStuff"
+import { getInputType } from "./inputs"
 import type { VuetifyInputProps } from "./InputProps"
 
 defineProps<VuetifyInputProps<From, Name>>()
