+++ packages/vue-components/src/components/OmegaForm/OmegaInternalInput.vue
@@ -27,7 +27,8 @@
 import { type DeepKeys, useStore } from "@tanstack/vue-form"
 import { computed, type ComputedRef, getCurrentInstance, useAttrs, useId, useSlots } from "vue"
 import type { InputProps, OmegaFieldInternalApi } from "./InputProps"
-import type { FieldValidators, MetaRecord, NestedKeyOf, TypeOverride } from "./OmegaFormStuff"
+import type { MetaRecord, NestedKeyOf } from "./meta/types"
+import type { FieldValidators, TypeOverride } from "./types"
 import OmegaInputVuetify from "./OmegaInputVuetify.vue"
 
 defineOptions({
