+++ packages/vue-components/src/components/OmegaForm/createUseFormWithCustomInput.ts
@@ -1,7 +1,7 @@
 import type { DeepKeys } from "@tanstack/vue-form"
 import { type Component, h } from "vue"
 import type { MergedInputProps } from "./InputProps"
-import { type DefaultTypeProps } from "./OmegaFormStuff"
+import { type DefaultTypeProps } from "./types"
 import OmegaInput from "./OmegaInput.vue"
 import { useOmegaForm } from "./useOmegaForm"
