+++ packages/vue-components/src/components/OmegaForm/index.ts
@@ -1,5 +1,56 @@
-export * from "./OmegaFormStuff"
-export { type OmegaConfig, type OmegaFormReturn, useOmegaForm } from "./useOmegaForm"
+export type {
+  BaseFieldMeta,
+  BooleanFieldMeta,
+  DateFieldMeta,
+  FieldMeta,
+  MetaRecord,
+  MultipleFieldMeta,
+  NestedKeyOf,
+  NumberFieldMeta,
+  SelectFieldMeta,
+  StringFieldMeta,
+  UnknownFieldMeta
+} from "./meta/types"
+export type {
+  BaseProps,
+  DefaultTypeProps,
+  FieldPath,
+  FieldPath_,
+  FieldValidators,
+  FormComponent,
+  FormProps,
+  FormType,
+  OmegaArrayProps,
+  OmegaAutoGenMeta,
+  OmegaError,
+  OmegaFormApi,
+  OmegaFormParams,
+  OmegaFormState,
+  OmegaInputProps,
+  OmegaInputPropsBase,
+  PrefixFromDepth,
+  TypeOverride,
+  TypesWithOptions
+} from "./types"
+export {
+  createMeta,
+  generateMetaFromSchema,
+  isNullableOrUndefined,
+  metadataFromAst
+} from "./meta/createMeta"
+export type { CreateMeta, FilterItems } from "./meta/createMeta"
+export { defaultsValueFromSchema } from "./meta/defaults"
+export { toFormSchema } from "./meta/redacted"
+export {
+  generateInputStandardSchemaFromFieldMeta,
+  makeStandardSchemaV1Hooks,
+  toLocalizedStandardSchemaV1
+} from "./validation/localized"
+export { deepMerge } from "./persistency"
+export { getInputType, type SupportedInputs } from "./inputs"
+
+export { FormErrors, OmegaFormKey, useErrorLabel, useOmegaForm } from "./useOmegaForm"
+export type { defaultValuesPriorityUnion, OF, OmegaConfig, OmegaFormReturn, Policies } from "./useOmegaForm"
 
 export { type ExtractTagValue, type ExtractUnionBranch, type InputProps, type MergedInputProps, type TaggedUnionOption, type TaggedUnionOptionsArray, type TaggedUnionProps } from "./InputProps"
 export { default as OmegaInput } from "./OmegaInput.vue"
@@ -8,7 +59,5 @@ export { default as OmegaTaggedUnion } from "./OmegaTaggedUnion.vue"
 export { default as OmegaTaggedUnionInternal } from "./OmegaTaggedUnionInternal.vue"
 
 export { useOnClose, usePreventClose } from "./blockDialog"
-
-export { getInputType } from "./OmegaFormStuff"
 
 export { createUseFormWithCustomInput } from "./createUseFormWithCustomInput"
