+++ packages/vue-components/src/components/OmegaForm/useOmegaForm.ts
@@ -9,13 +9,15 @@ import { fHoc } from "./hocs"
 import OmegaArray from "./OmegaArray.vue"
 import OmegaAutoGen from "./OmegaAutoGen.vue"
 import OmegaErrorsInternal from "./OmegaErrorsInternal.vue"
-import { defaultsValueFromSchema, DefaultTypeProps, type FormProps, generateMetaFromSchema, type OmegaFormApi, toFormSchema } from "./OmegaFormStuff"
+import { defaultsValueFromSchema } from "./meta/defaults"
+import { generateMetaFromSchema } from "./meta/createMeta"
+import { toFormSchema } from "./meta/redacted"
 import OmegaInput from "./OmegaInput.vue"
 import OmegaTaggedUnion from "./OmegaTaggedUnion.vue"
 import OmegaForm from "./OmegaWrapper.vue"
 import { usePersistency } from "./persistency"
 import { FormErrors, makeSubmitHandlers, wrapOnSubmit } from "./submit"
-import type { OF, OmegaConfig, OmegaFormReturn } from "./types"
+import type { DefaultTypeProps, FormProps, OF, OmegaConfig, OmegaFormApi, OmegaFormReturn } from "./types"
 
 import { makeRunPromise } from "@effect-app/vue/runtime"
